RAG_service/requirements.txt
mistralai==1.5.1
fastapi==0.115.11
uvicorn==0.34.0
requests==2.32.3

boto3==1.38.23
faiss-cpu==1.11.0
sentence-transformers==4.1.0
python-docx==1.1.2
PyMuPDF==1.25.5